requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
groq
python-multipart
openpyxl
